Von: "Michael Kerrisk" <mtk-manpages@gmx.net> An: Olivier Croquette <ocroquette@free.fr> Betreff: Re: 2.6.12 and setitimer Datum: Mon, 4 Jul 2005 08:36:35 +0200 (MEST)
Hi Olivier,

> You will probably consider adding also a note to point out that the bug 
> will stay a known bug of the 2.4 serie:
> 
> http://lkml.org/lkml/2005/7/1/165

First off, I _very_ much appreciate the fact that you keep 
informing me of the progress of this bug!  Thank you.

At the moment, I'm inlined yo leave the manual page as it is.
It currently reads:

       On certain systems (including x86), Linux ker&#8208;
       nels  before  version  2.6.12 have a bug which
       will produce premature timer expirations of up
       to  one  jiffy under some circumstances.  This
       bug is fixed in kernel 2.6.12.

To me that implies that the bug also affects kernels before 
2.4 -- e.g., 2.4.x.  Now, what would be interesting is if the
bug *does* get fixed in 2.4, then I could also add a note 
about the 2.4.x version where it is fixed.

In the meantime, I have added a note to myself (i.e., a comment
in the man page source) about this point.

If the bug *does* eventually get fixed in 2.4.x, and you 
hear of it, please do let me know.

Thanks,

Michael