execve.2: Linux now imposes a floor on the ARG_MAX limit
Starting with Linux 2.6.23, the ARG_MAX limit became settable via
(1/4 of) RLIMIT_STACK.  This broke ABI compatibility if RLIMIT_STACK
was set such that ARG_MAX was < 32 pages.  Document the fact that
since 2.6.25 Linux imposes a floor on ARG_MAX, so that the old limit
of 32 pages is guaranteed.

For some background on the changes to ARG_MAX in kernels 2.6.23 and
2.6.25, see:
    http://sourceware.org/bugzilla/show_bug.cgi?id=5786
    http://bugzilla.kernel.org/show_bug.cgi?id=10095
    http://thread.gmane.org/gmane.linux.kernel/646709/focus=648101,
    checked into 2.6.25 as commit a64e715fc74b1a7dcc5944f848acc38b2c4d4ee2.

Also some reordering/rewording of the discussion of ARG_MAX.

Signed-off-by: Michael Kerrisk <mtk.manpages@gmail.com>