open.2: minor: Fix O_CLOEXEC wording re fcntl() calls
The current wording suggests that only a single fcntl()
operation is needed to set the FD_CLOEXEC flag, when "proper"
usage would be fcntl(F_GETFD) + fcntl(F_SETFD) to get the
flags and then update them.  So change the wording to indicate
that more than one fcntl() operation is required.

Signed-off-by: Michael Kerrisk <mtk.manpages@gmail.com>