pthreads.7: Rework, and fix small error in, thread-safe function list
Integrate the changes that occurred in POSIX.1-2008 into the
main list (to be consistent with the list, elsewhere on this
page, of functions that are cancellation points).

Also, fix an error that said that strerror() was added to
the list in POSIX.1-2008.  It was strsignal() that was
added.  (strerror() was already in the list in POSIX.1-2001.)

Signed-off-by: Michael Kerrisk <mtk.manpages@gmail.com>