inotify.7: Fix description of max_user_watches
It seems that inotify(7) is wrong here:

"/proc/sys/fs/inotify/max_user_watches
    This specifies a limit on the number of watches that can be
    associated with each inotify instance."

On my system, the default value for this variable is 8192. But I
cannot create more than 8192 watches in total for the same UID
even when they are on different inotify instances. So I suggest
to rephrase this as: "This specifies an upper limit on the
number of watches that can be created per real user ID."

Signed-off-by: Michael Kerrisk <mtk.manpages@gmail.com>