mmap.2: Loosen language around how 'addr' hint is interpreted
Mel Gorman reported that in Linux 2.6.27, 'addr' is rounded
down to a page boundary.

Before kernel 2.6.26, if 'addr' was taken as a hint, it was
rounded up to the next page boundary.  Since Linux 2.6.24,
it is rounded down.  Therefore, loosen the description of
this point to say that the address is rounded to "a nearby
page boundary".

Signed-off-by: Michael Kerrisk <mtk.manpages@gmail.com>
Reported-by: Mel Gorman <mel@csn.ul.ie>