nscd.8: Correct NOTES section on reloading configuration files
It behaved this way at least since "Sun Oct 18 15:02:11 1998 +0000",
some four months after including the nscd implementation in glibc. But
there does seem to be a short window between glibc-2.1 and glibc-2.1.3
when nscd -i was not available, I don't think it's worth muddling the
point of the page with that, though.

Signed-off-by: Petr Baudis <pasky@suse.cz>
Signed-off-by: Michael Kerrisk <mtk.manpages@gmail.com>