splice.2: ERRORS: Add EINVAL case for file opened O_APPEND
Target file cannot be opened in append (O_APPEND) mode

In kernels prior to v2.6.27 splice() to a file in append mode is broken,
and since that version it is disallowed. It is possible this behaviour
may change in the future; see the kernel commit message
(efc968d450e013049a662d22727cf132618dcb2f) for more information.

Signed-off-by: Mark Hills <mark@pogo.org.uk>
Signed-off-by: Michael Kerrisk <mtk.manpages@gmail.com>