accept.2, read.2, recv.2, send.2, write.2: Fix discussion of EAGAIN/EWOULDBLOCK errors
For a non-blocking socket, POSIX.1-2001/2008 allow either
EAGAIN or EWOULDBLOCK to be returned in cases where a call
would have blocked.  Although these constants are defined
with the same value on most Linux architectures (PA-RISC
is the exception), POSIX.1 does not require them to have
the same value.  Therefore, a portable application using
the sockets API should test for both errors when checking
this case.

(NB POSIX.1 only mentions EWOULDBLOCK in the context of
the sockets interfaces.)

Change made after a note cross-posted on linux-arch@vger,
http://thread.gmane.org/gmane.linux.debian.ports.hppa/5615
and a suggestion for write(2) from Carlos O'Donell

Reported-by: Carlos O'Donell <carlos@systemhalted.org>
Signed-off-by: Michael Kerrisk <mtk.manpages@gmail.com>