fork.2: Document fork() behaviour for the Linux native AIO io_context
It was noted on lkml that the fork behaviour is documented
for the POSIX AIO calls, but not for the Linux native calls.
Here is a patch which adds a small blurb that folks will
hopefully find useful.

Upon fork(), the child process does not inherit the
io_context_t data structures returned by io_setup,
and thus cannot submit further asynchronous I/O or
reap event completions for said contexts.

Signed-off-by: Jeff Moyer <jmoyer@redhat.com>
Signed-off-by: Michael Kerrisk <mtk.manpages@gmail.com>