koi8-r.7: Fix SEE ALSO reference and letter names
he koi8-r(7) (Russian Net Character Set) manual page refers to
iso-8859-7(7) manual page, which is the Latin/Greek one. I guess it
should refer instead to the iso-8859-5(7) (Latin/Cyrillic) one. This is
addressed at the end of the patch.

It has also been spotted that letter names are different in this manual
page and in the Unicode related page [0], the first part of the page
address this.

        0: http://www.unicode.org/charts/PDF/U0400.pdf

Signed-off-by: Michael Kerrisk <mtk.manpages@gmail.com>