kexec_load.2: Additions and edits by mtk
Various wording and layout improvements.
Fixed the name of a constant: s/KEXEC_ARCH_I386/KEXEC_ARCH_386/.
Added RETURN VALUE and ERRORS sections.
Removed details of using syscall; the reader can find them in
syscall(2).
Added some details for KEXEC_PRESERVE_CONTEXT.
Revised the text mentioning the kernel header, since it is
not yet exported, and it's not certain that it will be.

Signed-off-by: Michael Kerrisk <mtk.manpages@gmail.com>