err.3: Clean up description of error message source
In the second paragraph of DESCRIPTION section, one of the
sources of error messages is incorrect: the four functions obtain
error message only from errno, and "a code" is just relevant
with errc() and warnc(), which are not present on Linux.
see http://www.unix.com/man-page/freebsd/3/ERR/ .

Then, the third paragraph becomes a duplicate.

Signed-off-by: Michael Kerrisk <mtk.manpages@gmail.com>