unlink.2: Improve EBUSY description
The current version of the page says (under ERRORS):

> EBUSY (not on Linux)
>  The  file pathname cannot be unlinked because it is being
>  used by the system or another process and the
>  implementation considers this an error.

But if you look in the kernel source file fs/namei.c
at routine may_delete() you'll see two occasions where
an EBUSY is generated. So the above "not on Linux" is wrong.

I suggest that this '(not on Linux)' be removed. It may
also improve the page if the commentary is reworded to give a
better description of the two situations. Something along
the lines of:
> The file pathname cannot be unlinked because it is being used
> by the system, e.g. if some (relative) rootdir is mounted upon it,
> or because the NFS client software created it to represent an
> active but otherwise nameless inode ("NFS silly renamed").

Just FYI: the NFS silly rename (you'll find this term mentioned
in the kernel source code) is described at:
       http://nfs.sourceforge.net/#section_d
under section D2. The reason I found this manpage bug is because
I stumbled upon one of these silly renamed files, and an strace
of the rm-command revealed the EBUSY return errno.

Signed-off-by: Michael Kerrisk <mtk.manpages@gmail.com>