nsswitch.conf.5: Significant rewrites and improvements
This patch applies to nsswitch.conf.5 in man-pages-3.36.

My changes almost completely rewrite large sections of the
man page. They are needed to add clarity, correct grammar,
reduce confusion, and bring up-to-date with the latest glibc.
I have checked the man page against the nss source code in
glibc 2.14.90.

Historical notes are demoted to the footer.

The rewrite makes the man page much clearer to
understand, more authoratitive, and easier to read.

Signed-off-by: Michael Kerrisk <mtk.manpages@gmail.com>