strchr.3: Explain behavior when seraching for '\0'
The strchr(3) page does `not explain the behavior if the
character to search for is specified as a null character ('\0').
According to my copy of Harbison and Steele, since the terminator
is considered part of the string, a call such as:

  strchr("hello", '\0')

will return the address of the terminating null in the specified
string.

strchr(3) is inconsistent with index(3) which states:

    "The terminating null byte is considered to be
    a part of the strings."

Adding such a note to strchr(3) is also important since it is not
unreasonable to assume that strchr() will return NULL in this
scenario. This leads to code like the following which is
guaranteed to fail should get_a_char() return '\0':

    char string[] = "hello, world";
    int c = get_a_char();

    if (! strchr(string, c))
        fprintf(stderr, "failed to find character in string\n");

Signed-off-by: Michael Kerrisk <mtk.manpages@gmail.com>