utf-8.7: Two clarifications
This patch clarifies that 0xc0 and 0xc1 are not valid in any UTF-8
encoding[0], and it also references RFC 3629 instead of RFC 2279.

[0] In order to have 0xc0, you'd have to have a two-byte encoding
with all the data bits zero in the first byte (and thus only six
bits of data), which would be an ASCII character encoded in the
non-shortest form.  Similarly with 0xc1.

See http://bugs.debian.org/cgi-bin/bugreport.cgi?bug=538641

Signed-off-by: Michael Kerrisk <mtk.manpages@gmail.com>