icmp.7, ipv6.7, packet.7, raw.7, rtnetlink.7, unix.7, x25.7: Remove names of constants from NAME line
Some of the sockets/network protocol pages included names of
the corresponding address family constants in the NAME line,
but this wasn't done consistently across all pages, and probably
it adds little value in those pages that did do this. So, remove
these constants from those pages that have them in the NAME
section.

Signed-off-by: Michael Kerrisk <mtk.manpages@gmail.com>