iso_8859-1.7: Explanation of SOFT HYPHEN and the code for it
<groff: iso_8859-2.7>:89: warning: can't find special character `shc'

This is the only "iso_8859-*.7" file that has this (now)
undefined character.  The code in column four in "iso_8859-1.7" is
"0x2D" ("HYPHEN, MINUS SIGN" or "HYPHEN-MINUS") instead of "0xAD".
See Debian bug 156154 (or package "manpages").

There should be an explanation for this graphic character and the
code should be 0xAD in iso_8859-1.7 (as in all others), even
though "[gn]roff" does not display a "HYPHEN" in that position of
the table.

The line with "SOFT HYPHEN" gets a footnote and a short
explanation.

Signed-off-by: Michael Kerrisk <mtk.manpages@gmail.com>