epoll_wait.2: Describe timeout limitation in kernels < 2.6.37
As reported by Fredrik (and as far as I can tell the problem
went back to 2.6.0):

The timeout argument has an upper limit. Any values above that
limit are treated the same as -1, i.e. to wait indefinitely.
The limit is given by:

    #define EP_MAX_MSTIMEO min(1000ULL * MAX_SCHEDULE_TIMEOUT / HZ, \
	(LONG_MAX - 999ULL) / HZ)

That is, the limit depends on the size of a long and the timer
frequency.  Assuming the a long is never smaller than 32 bits
and HZ never larger than 1000, the worst case is 35 minutes.
I think this should be mentioned under "BUGS".

Although this is likely to be fixed in the future
(http://lkml.org/lkml/2010/8/8/144), the problem exists in
at least 2.6.14 - 2.6.35. I don't know if select(2) and poll(2)
are affected.

https://bugzilla.kernel.org/show_bug.cgi?id=20762

Reported-by: Fredrik Arnerup <arnerup@kth.se>
Signed-off-by: Michael Kerrisk <mtk.manpages@gmail.com>