link.2, mkdir.2, mknod.2, open.2, rename.2, symlink.2, write.2, mkfifo.3: Document EDQUOT error
The return error EDQUOT is not documented in open(2), write(2),
symlink(2) etc..

Whether inodes or disk blocks are required for each function
is something I based on received wisdom and BSD documentation,
rather than tracing the code to the kernel. For symlink(2)
this certainly depends on the file system type.

This patch was previously sent to the linux-man mailing list
as an RFC in September 2010.

Signed-off-by: Mark Hills <mark@pogo.org.uk>
Signed-off-by: Michael Kerrisk <mtk.manpages@gmail.com>