getxattr.2, removexattr.2, setxattr.2: Note that ENOATTR is a synonym for ENODATA
Various people have pointed out that strace(1) shows ENODATA
for the case where the named attribute does not exist, or
the process has no access to this attribute. ENODATA
and ENOATTR are in fact defined as synonyms. Point this out
in the man page, so that people understand the strace(1) info.

See https://bugzilla.kernel.org/show_bug.cgi?id=51871

Reported-by: Andrew Perepechko <andrew_perepechko@xyratex.com>
Reported-by: <starlight.2013q1@binnacle.cx>
Signed-off-by: Michael Kerrisk <mtk.manpages@gmail.com>