random.4: Document write and document the ioctl interface of /dev/random
The update consists out of two parts:
 - a minor thing which just documents what happens if a write to
   /dev/(u)random is performed, which is used in the example
   script but not explicitly mentioned.
 - the other (biggest) part is the documentation of the ioctl()
   interface which /dev/(u)random exposes. This ioctl() lives in
   drivers/char/random.c and the primitives can be found in
   include/linux/random.h

One comment remains, there used to be an RNDGETPOOL ioctl() which
disappeared in v2.6.9. I found two patches on the net:
- http://www.kernel.org/pub/linux/kernel/people/akpm/patches/2.6/2.6.8.1/2.6.8.1-mm4/broken-out/dev-random-remove-rndgetpool-ioctl.patch
- https://lkml.org/lkml/2004/3/25/168

But as far as I can tell the first one got applied but the 2nd
one seems more correct. The result is that even today one can
still find traces of the RNDGETPOOL ioctl() in the header files.
Is this there for historical reasons or because it might break
userspace, even though using it will just give an EINVAL.

Signed-off-by: Michael Kerrisk <mtk.manpages@gmail.com>