readir.3, seekdir.3, telldir.3: Eliminate the implication that these functions deal with "offsets"
The directory position dealt with by the readdir() and
friends is not a simple file offset in modern file systems.
Typically, it is some kind of cookie value. Add text and
make other changes to these pages to eliminate the
implication that this is an offset, and warn the reader
that directory positions should be treated strictly as
opaque values.

In the process, rename the 'offset' argument of seekdir(3)
to 'loc', and add some text to readdir(3) to note that
the 'd_off' field is the same value returned by telldir(3)
at the current directory position.

See also https://lwn.net/Articles/544298/

Signed-off-by: Michael Kerrisk <mtk.manpages@gmail.com>