elf.5: Add byte positions for all EI_xxx fields
When describing e_ident, most of the EI_xxx defines mention the
exact byte number.  This is useful when manually hacking an ELF
with a hex editor.  However, the last few fields don't do this,
which means you have to count things up yourself.
Add a single word to each so you don't have to do that.

Signed-off-by: Mike Frysinger <vapier@gentoo.org>
Signed-off-by: Michael Kerrisk <mtk.manpages@gmail.com>