io_setup.2: Clarify nr_events
nr_events is technically the number of completion events that can
be stored in the completion ring.  The wording of the man page:
"capable of receiving at least nr_events" seems dubious to me,
only because I worry that folks might interpret that to mean
'nr_events' total, instead of 'nr_events' concurrently.

Further, I've added information on where to find the per-user
limit on 'nr_events', /proc/sys/fs/aio-max-nr.  Let me know if
you think that is not relevant.

Signed-off-by: Jeff Moyer <jmoyer@redhat.com>
Signed-off-by: Michael Kerrisk <mtk.manpages@gmail.com>