ip.7: Document IP_MULTICAST_ALL
This commit adds documentation for the IP_MULTICAST_ALL socket
option.

The option was added to the Linux kernel in 2.6.31:

    Author Nivedita Singhvi <niv@us.ibm.com>
    Commit f771bef98004d9d141b085d987a77d06669d4f4f

The description is based on a previous one [3] posted by the
original author of the code -- Nivedita, but it is slightly
re-worded.

I tested it myself and it works as described.

References:
    [1] http://lxr.free-electrons.com/source/net/ipv4/ip_sockglue.c#L972
    [2] http://lxr.free-electrons.com/source/net/ipv4/igmp.c#L2267
    [3] http://patchwork.ozlabs.org/patch/28902/

Signed-off-by: Radek Pazdera <rpazdera@redhat.com>
Signed-off-by: Michael Kerrisk <mtk.manpages@gmail.com>