io_setup.2: Clarify the nr_events parameter
Currently the io_setup.2 man page describes what the kernel really
does, i.e., that the resulting context may be able to hold more
than the 'nr_event's operations because the memory allocated in
kernel is rounded to be multiple of page size.

It is better not to expose this implementation detail and
simply state that the resulting context is suitable for
'nr_events' operations.

Signed-off-by: Michael Kerrisk <mtk.manpages@gmail.com>
Signed-off-by: Cyril Hrubis <chrubis@suse.cz>
Acked-by: Jeff Moyer <jmoyer@redhat.com>